[DPR2-1514] Make async beans optional
